net: move protocol state struct into its own module
This is a purely mechanical change, as part of continuously grooming the
scrolls.
