+++ librad/src/net/protocol.rs
@@ -7,7 +7,6 @@ use std::{
     fmt::Debug,
     future::Future,
     net::SocketAddr,
-    ops::Deref,
     pin::Pin,
     sync::Arc,
     task::{Context, Poll},
@@ -16,12 +15,10 @@ use std::{
 
 use futures::{
     channel::mpsc,
-    future::{BoxFuture, FutureExt as _, TryFutureExt as _},
+    future::{BoxFuture, FutureExt as _},
     stream::StreamExt as _,
 };
-use governor::{Quota, RateLimiter};
 use nonempty::NonEmpty;
-use nonzero_ext::nonzero;
 use rand_pcg::Pcg64Mcg;
 use tracing::Instrument as _;
 
@@ -34,12 +31,9 @@ use super::{
 use crate::{
     git::{
         self,
-        p2p::{
-            server::GitServer,
-            transport::{GitStream, GitStreamFactory},
-        },
+        p2p::{server::GitServer, transport::GitStreamFactory},
         replication,
-        storage::{self, PoolError, PooledRef},
+        storage,
     },
     paths::Paths,
     signer::Signer,
@@ -66,6 +60,9 @@ mod tick;
 mod tincans;
 pub(super) use tincans::TinCans;
 pub use tincans::{Interrogation, RecvError};
+
+mod state;
+use state::{State, StateConfig, Storage};
 
 #[derive(Clone, Debug)]
 pub struct Config {
@@ -266,147 +263,6 @@ pub trait ProtocolStorage<A>: broadcast::LocalStorage<A> + storage::Pooled + Sen
 impl<A, T> ProtocolStorage<A> for T where
     T: broadcast::LocalStorage<A> + storage::Pooled + Send + Sync
 {
-}
-
-type Limiter = governor::RateLimiter<
-    governor::state::direct::NotKeyed,
-    governor::state::InMemoryState,
-    governor::clock::DefaultClock,
->;
-
-#[derive(Clone)]
-struct Storage<S> {
-    inner: S,
-    limiter: Arc<Limiter>,
-}
-
-impl<S> From<S> for Storage<S> {
-    fn from(inner: S) -> Self {
-        Self {
-            inner,
-            limiter: Arc::new(RateLimiter::direct(Quota::per_second(
-                // TODO: make this an "advanced" config
-                nonzero!(5u32),
-            ))),
-        }
-    }
-}
-
-impl<S> Deref for Storage<S> {
-    type Target = S;
-
-    fn deref(&self) -> &Self::Target {
-        &self.inner
-    }
-}
-
-#[async_trait]
-impl<A, S> broadcast::LocalStorage<A> for Storage<S>
-where
-    A: 'static,
-    S: broadcast::LocalStorage<A>,
-    S::Update: Send,
-{
-    type Update = S::Update;
-
-    async fn put<P>(&self, provider: P, has: Self::Update) -> broadcast::PutResult<Self::Update>
-    where
-        P: Into<(PeerId, Vec<A>)> + Send,
-    {
-        self.inner.put(provider, has).await
-    }
-
-    async fn ask(&self, want: Self::Update) -> bool {
-        self.inner.ask(want).await
-    }
-}
-
-impl<S> broadcast::ErrorRateLimited for Storage<S> {
-    fn is_error_rate_limit_breached(&self) -> bool {
-        self.limiter.check().is_err()
-    }
-}
-
-#[async_trait]
-impl<S> storage::Pooled for Storage<S>
-where
-    S: storage::Pooled + Send + Sync,
-{
-    async fn get(&self) -> Result<PooledRef, PoolError> {
-        self.inner.get().await
-    }
-}
-
-#[derive(Clone, Copy)]
-pub struct StateConfig {
-    pub replication: replication::Config,
-    pub fetch: config::Fetch,
-}
-
-#[derive(Clone)]
-struct State<S> {
-    local_id: PeerId,
-    endpoint: quic::Endpoint,
-    git: GitServer,
-    membership: membership::Hpv<Pcg64Mcg, SocketAddr>,
-    storage: Storage<S>,
-    phone: TinCans,
-    config: StateConfig,
-    nonces: nonce::NonceBag,
-    caches: cache::Caches,
-}
-
-#[async_trait]
-impl<S> GitStreamFactory for State<S>
-where
-    S: ProtocolStorage<SocketAddr, Update = gossip::Payload> + Clone + 'static,
-{
-    async fn open_stream(
-        &self,
-        to: &PeerId,
-        addr_hints: &[SocketAddr],
-    ) -> Option<Box<dyn GitStream>> {
-        let span = tracing::info_span!("open-git-stream", remote_id = %to);
-
-        let may_conn = match self.endpoint.get_connection(*to) {
-            Some(conn) => Some(conn),
-            None => {
-                let addr_hints = addr_hints.iter().copied().collect::<Vec<_>>();
-                io::connect(&self.endpoint, *to, addr_hints)
-                    .instrument(span.clone())
-                    .await
-                    .map(|(conn, ingress)| {
-                        tokio::spawn(
-                            io::streams::incoming(self.clone(), ingress).instrument(span.clone()),
-                        );
-                        conn
-                    })
-            },
-        };
-
-        match may_conn {
-            None => {
-                span.in_scope(|| tracing::error!("unable to obtain connection"));
-                None
-            },
-
-            Some(conn) => {
-                let stream = conn
-                    .open_bidi()
-                    .inspect_err(|e| tracing::error!(err = ?e, "unable to open stream"))
-                    .instrument(span.clone())
-                    .await
-                    .ok()?;
-                let upgraded = upgrade::upgrade(stream, upgrade::Git)
-                    .inspect_err(|e| tracing::error!(err = ?e, "unable to upgrade stream"))
-                    .instrument(span)
-                    .await
-                    .ok()?;
-
-                Some(Box::new(upgraded))
-            },
-        }
-    }
 }
 
 impl<R, A> broadcast::Membership for membership::Hpv<R, A>
